requirements.txt
fastmcp
streamlit
pypdf
python-dotenv
langchain-chroma
chromadb
langchain
langchain-community
langchain-google-genai
google-generativeai